+++ src/xml/xml.cc
@@ -406,5 +406,11 @@ std::string WriteXML(const mjSpec* spec, char* error, int nerror) {
 
   mjXWriter writer;
   writer.SetModel(spec);
-  return writer.Write(error, nerror);
+
+  try {
+    return writer.Write(error, nerror);
+  } catch (mjXError err) {
+    mjCopyError(error, err.message, nerror);
+    return "";
+  }
 }